Add Prometheus integration, update TravisCI (#212)
* Add Prometheus integration, update TravisCI

This adds Prometheus integration via the --prometheus flag. The flags
defaults to empty. When the flag is non-empty it launches an HTTP server
at the specified endpoint (e.g. "localhost:8080"), that exposes the
Prometheus metric source API at the /metrics path.

This also updates TravisCI to download dependencies more sanely, and
bumps the version of Golang used by TravisCI to 1.7.4.

* Clarify Prometheus option help text