Refork net/http to resolve memory leak (#268)
This reforks the Golang net/http library to resolve the memory leak in zgrab (#187). This also reapplies all necessary modifications to the Golang http library for it to interface with zgrab.

HTTP/2 is supported in this reforked library however there were compatibility issues with zcrypto/tls and crypto/tls that caused tests to fail. Since it's not the goal of this PR, I've commented out relevant HTTP/2 tests. This will allow us to support it at a later date without another fork of the stdlib (#270)

Closes #187
Closes #266