Pedantic Fixes (#272)
Very pedantic changes. 

Adds Go Report Card to README, useful because A) it runs gofmt, go vet, go lint, spelling errors, and checks for Ineffectual Assignments and B) because it says we have an A+ on our code. (Really the main use here is that it provides a single click to check all the pedanticness of the code)