Modification of zgrab to take advantage of new functionality in zcrypto
I've added InsecureNoValidation to the zcrypto/tls package.  This is
like CertsOnly in that it's interested in grabbing raw certs for parsing
and processing offline, but is 'kinder' because it finishes the
handshake.  Therefore, it's also a bit slower.

My initial testing showed --tls-no-validation to be about 2x slower than
--tls-certs-only (however, --tls-certs-only is about 10 times faster
than the default, so --tls-no-validation is still significantly faster)

Hitting a host near my test box 10 times in a row and averaging times:

--tls-skip-validation
.05530s

--tls-certs-only
.01820s

default
.24770s